Backend/requirements.txt
# Core
pydantic>=2.0,<3.0
langgraph>=0.2,<1.0
langchain-core>=0.2,<1.0

# LLM / SDK
httpx>=0.27,<1.0
cerebras-cloud-sdk>=1.0.0

# Embeddings
sentence-transformers>=2.7,<4.0
torch>=2.2,<3.0     # CPU build; install torchvision separately if needed

# Vector store
faiss-cpu>=1.7,<2.0

# Persistence
# sqlite3 is stdlib — no install required

# Utilities
numpy>=1.24,<2.0
typing-extensions>=4.9
python-dotenv>=1.0,<2.0

# API Server
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
websockets>=12.0,<14.0

black